--- jsr.jsonc
@@ -1,5 +1,0 @@
-{
-    "name": "@cross/deepmerge",
-    "version": "0.1.1",
-    "exports": "./mod.ts"
-}
